[Refactor]: convertible의 receviceCurrentDate 프로퍼티 코드 이동
